Added talk bubbles and object to the game
Idle Volfram sprite added to the game
Idle Liel sprite added to the game
Toybox sprite added to the game
Talk bubble sprite added to the game.

Talk bubble will appear when Enfys is near the other objects. Current
code is just a placeholder that will later be replaced with
"Object.State" that will be linked to a specific object key for
interacting.